/trunk/core: - Fix pagan zone parsing - Fix clan hall deco seen only after entering clan hall - Added a config to remove npc skill verification warnings
git-svn-id: http://svn.l2jfree.com/l2jfree@7544 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/model/actor/instance/L2BirthdayHelperInstance.java
@@ -29,7 +29,7 @@ import com.l2jfree.gameserver.templates.chars.L2NpcTemplate;
  */
 public final class L2BirthdayHelperInstance extends L2Npc
 {
-	private static final long HELPING_LENGTH = 1200000;
+	private static final int HELPING_LENGTH = 1200000;
 	private static final int BIRTHDAY_CAKE_EFFECT = 22035;
 
 	private int _ownerId;
